pyproject.toml
[project]
name = "api-project"
version = "0.1.0"
description = ""
authors = [
    {name = "verkonia",email = "v.azmukova@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "uvicorn (>=0.34.2,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "db (>=0.1.1,<0.2.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic (>=2.11.3,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "faker (>=37.1.0,<38.0.0)",
    "pylint (>=3.3.6,<4.0.0)"
]

[tool.poetry]
packages = [{include = "api_project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.MASTER]
init-hook = '''
import sys, pathlib;
sys.path.append(str(pathlib.Path(__file__).parent.resolve()));
sys.path.append(str(pathlib.Path(__file__).parent.joinpath("app").resolve()));
'''

[tool.pylint.IMPORTS]
ignored-modules = [
    "app.dbs.database",
    "app.routes",
    "app.dbs.session",
    "app.models.models",
    "app.config",
    'app.schemas.schemas',
    'app.core.security'
]